db_app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
prometheus-client==0.20.0
pydantic==2.8.2
python-dotenv==1.0.1
pika==1.3.2
python-jose[cryptography]==3.3.0
